backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
psycopg2-binary==2.9.9

# Environment variables
python-dotenv==1.0.0

# Data validation
pydantic==2.5.3
pydantic-settings==2.1.0

# LLM for clustering (KPI #2)
groq==1.0.0

# Sentiment analysis (KPI #13)
vaderSentiment==3.3.2
